feat: move request processing stuff to server package
